requirements.txt
yt-dlp
pydantic
ffmpeg-python
sentence-transformers
